add DashBoardCard
