+++ app/(pages)/dashboard/page.tsx
@@ -2,7 +2,7 @@
 import styles from "./page.module.css";
 import { Center, Container, Heading, Spinner } from "@chakra-ui/react";
 import { useRequireAccount } from "@/app/hooks/common/useRequireAccount";
-import DashBoardList from "@/app/components/airdrops/DashBoardList";
+import DashBoardList from "@/app/components/dashboard/DashBoardList";
 import Test from "./test";
 
 export default function Dashboard() {
